UI full rebuild
Ui out of moduleEditor.
Now working with objects inside src/Editor
