+++ Engine/Engine.vcxproj
@@ -140,6 +140,13 @@
     <ClInclude Include="src\Application.h" />
     <ClInclude Include="src\Component.h" />
     <ClInclude Include="src\DebugLeaks.h" />
+    <ClInclude Include="src\EditorAbout.h" />
+    <ClInclude Include="src\EditorComponent.h" />
+    <ClInclude Include="src\EditorConfig.h" />
+    <ClInclude Include="src\EditorConsole.h" />
+    <ClInclude Include="src\EditorMainMenu.h" />
+    <ClInclude Include="src\EditorPlay.h" />
+    <ClInclude Include="src\EditorViewport.h" />
     <ClInclude Include="src\GameObject.h" />
     <ClInclude Include="src\Globals.h" />
     <ClInclude Include="src\Libraries\Assimp\include\assimp\aabb.h" />
@@ -331,6 +338,12 @@
   <ItemGroup>
     <ClCompile Include="src\Application.cpp" />
     <ClCompile Include="src\Component.cpp" />
+    <ClCompile Include="src\EditorAbout.cpp" />
+    <ClCompile Include="src\EditorConfig.cpp" />
+    <ClCompile Include="src\EditorConsole.cpp" />
+    <ClCompile Include="src\EditorMainMenu.cpp" />
+    <ClCompile Include="src\EditorPlay.cpp" />
+    <ClCompile Include="src\EditorViewport.cpp" />
     <ClCompile Include="src\GameObject.cpp" />
     <ClCompile Include="src\Libraries\ImGui\imgui.cpp" />
     <ClCompile Include="src\Libraries\ImGui\imgui_demo.cpp" />
