+++ Engine/Game/imgui.ini
@@ -19,13 +19,13 @@ Size=500,400
 Collapsed=0
 
 [Window][Console]
-Pos=0,781
-Size=1536,235
+Pos=0,788
+Size=1459,237
 Collapsed=0
 
 [Window][Configuration]
 Pos=0,19
-Size=384,762
+Size=364,769
 Collapsed=0
 
 [Window][Hello, world!]
@@ -39,28 +39,28 @@ Size=65,48
 Collapsed=0
 
 [Window][About]
-Pos=60,60
+Pos=552,94
 Size=744,412
 Collapsed=0
 
 [Window][GameObjectHierarchy]
-Pos=1536,527
-Size=384,489
+Pos=1459,532
+Size=364,494
 Collapsed=0
 
 [Window][Inspector]
-Pos=1536,19
-Size=384,508
+Pos=1459,19
+Size=364,513
 Collapsed=0
 
 [Window][PlayStopStep]
-Pos=889,21
-Size=178,64
+Pos=364,38
+Size=1094,57
 Collapsed=0
 
 [Window][Viewport]
-Pos=384,19
-Size=1152,762
+Pos=364,19
+Size=1094,769
 Collapsed=0
 
 [Docking][Data]
