+++ Engine/src/Editor.h
@@ -1,0 +1,15 @@
+#pragma once
+#include "Editor.h"
+#include "Libraries/MathGeoLib/MathGeoLib.h"
+#include <string>
+#include <vector>
+
+class Editor
+{
+public:
+	Editor();
+	virtual ~Editor();
+
+	virtual void Update() {};
+
+};
