+++ Engine/src/EditorAbout.h
@@ -1,0 +1,12 @@
+#pragma once
+#include "EditorComponent.h"
+
+class EditorAbout : public EditorComponent
+{
+public:
+    EditorAbout() { type = About; };
+    ~EditorAbout() {};
+
+    void Draw(int w, int h);
+};
+
