+++ Engine/src/EditorConsole.h
@@ -1,0 +1,17 @@
+#pragma once
+#include "EditorComponent.h"
+
+class EditorConsole : public EditorComponent
+{
+public:
+    EditorConsole() { type = Console; };
+    ~EditorConsole() {};
+
+    void Draw(int w, int h);
+
+
+private:
+
+
+    void ClearLog();
+};
